Important change in the procedure "move"
The original version of the "move" procedure has been restored: Now cows move to neighboring patches instead of teleporting to any patch in the world.
However, it must be emphasized that in this current version of the model, the amount of dry matter per hectare is large enough that cows don't need to move to another patch, even when the system reaches equilibrium (in a normal year (coefclima 1) and with an initial grass height of 7 cm)
Maybe I should try moving cows to neighboring patches when gh-individual is below 5?
Maybe this makes sense because cows maintain their live weight when the pasture is above 5 cm, so if a cow is sharing the same patch with other animals, if she is sharing the resource with other cows, and she notices that the gh-individual is below 5 cm, maybe she should move?.
If I decide to do this, I should move the "move" procedure below the "kgDM/cow" procedure because it is in this procedure that the cows calculate the "gh-individual" variable.