Fixed bug with cow creation when rotational strategy is selected
Change the code in the "setup-livestock" procedure related to the rules for creating cows when the rotational grazing strategy is selected.

With the current code, you cannot generate more cows than the total number of patches (because of the "n-of" and "sprout 1" primitives).

This code has been changed: you can now generate more cows than patches if needed.