1) Fixed a bug related to the 'Move' procedure; 2) Fixed the random location for the born-calf
1)
Cows can move into a white area (wall), because in the procedure "move" one of the conditions of the ifelse was "move to one of the neighbors".
This is now fixed by adding to this line "move-to-one-of-neighbors with [wall = 0]".

2)
Now born-calfs spawn in the same patch as the mother (I just have to remove the line "setxy random-pxcor random-pycor"