1) Fixed a bug that caused the system to collapse under open access strategy.; 2) Changed the units of the "STOP-SIMULATION-AT" slider; 3) Brief change in the way cows share the resource in the "kgDM/cow" procedure
1) This bug was caused by the new change I just made (removing the line of code "setxy random-pxcor random pycor" for born-calves).
For some reason, under the "open access" strategy, the system collapsed when born-calves stayed at the same patch as their mothers.
In order to solve this problem, I made the born-calves to move away from their mothers' patch by adding the next line: "move-to one-of neighbors with [wall = 0]"

2) I changed the units of the "STOP-SIMULATION-AT" slider from days to years.

3) Changed line from:

"ask cows-here [set DM-kg-cow DM-kg-ha / count cows-here]"

to:

"ask cows-here with [weaned-calf? or heifer? or steer? or cow? or cow-with-calf?] [set DM-kg-cow DM-kg-ha / count cows-here with [weaned-calf? or heifer? or steer? or cow? or cow-with-calf?] ]"

The reason for this change is that cows were counting born calves (which do not eat grass) when sharing the resource. With this change, the resource is shared between all classes that can eat grass, excluding born calves.