Updated the way in which BCS is calculated based on the recommendations of the authors.
The update has been made to the BCS plot and monitors.

One important thing to notice: this updated has been made based on the recommendations of the authors:

“El BCS se calcula siempre como (PV-220)/40 para todos los animales”
"BCS is always calculated as "(LW - 220)/40" for all animals"

But with this change, if an animal is below 220 kg of LW, the BCS is negative.

With this in mind, it makes a lot of sense what the author said:

“Se fija para todos los animales adultos el peso mínimo de supervivencia en 220 kg”
"All animals have a minimum weight of 220kg."