1) Walls removed; 2) World size will be always an even number; 3) Next steps
1) Walls (i.e. white patches where grass doesn't grow) removed.

2) Changes to world size configuration: World size will always be an even number (before it was always an odd number). This change is necessary to have four paddocks of the same size. To achieve this, I had to modify the parameters of the sliders "set-x-size" and "set-y-size":

- Minimum: 2

- Increment: 2

- Maximum: 100

3) The next step is to get the cows to move within the current paddock (also don't forget how the born calf is spawned!).

I think the key to achieving this is within the "move" procedure (obviously...). My idea is to put the first half of this procedure inside an "if (spatial-management = "open access")" condition, so the cows will only read this code if "open access" is selected. This will allow me to write new code (practically the same as "open access" but with some conditions) when rotational grazing is selected, to tell cows to "move to one of the neighbors with ["if you are in paddock X, move to a patch with paddock X"].