Added "DM-cm-ha?" chooser to UI
Added "DM-cm-ha?" chooser, allowing the user to choose between a DM-cm-ha value of 180 or 180/92.

This will serve as a temporary solution to the problem of static cows (because with a value of 180 cows won't move because there are a lot of resources in the system).
The resource decreases when we divide 180 / 92. It is called temporary because it is just a "random" number without much meaning. I need to replace it with a number that makes sense.