Changed lactation period from 184 to 246 days
Important remark about the "born-calf" and "weaned-calf" stages:

In the code, I have observed possible inconsistencies with the info given in the papers....

On the one hand, in the figure with UML diagram (page 13 document "Notes_SB") it says that the natural wean is at 8 months (246 days). From this, a priori, it should be deduced that, therefore, the duration of the "Born-calf" life cycle is 246 days (this deduction is supported by what is said that a born-calf gains 0.61 kg/day, and that the initial weight of weaned-calf is 150 kg. Therefore, 0.61 kg/day * 246 days = 150.06 kg. "The born calves do not depend on grasslands. We assume that born calves increase their live weight by 0.61 kg/day. After 6 months, they should reach 150 Kg (the initial weight for weaned calves". ATTENTION!!!! THIS "6 months" MUST BE AN ERRATA!!! BECAUSE IF WE DO THE CALCULATIONS: 6 months = 184 days. 0.61 kg/day * 184 days = 112.24 kg. DOES NOT RESULTS IN 150 kg.

Because of this, I have changed the lactation period from 184 o 256 days