1) Substantial changes in the logic of how to implement different climate change scenarios; 2) Added a “Go (1 year)” button in the interface;  3) Changed default value for the “DM-cm-ha” variable
1) Substantial changes in the logic of how to implement different climate change scenarios:
o	Yesterday, I updated the model with three new variables that will allow us to simulate different climate change scenarios. These variables are “season-length”, “season-climacoef” and “season-climacoef-distribution”. The logic is that each day, a new climacoef will be calculated depending on the season and the seasonal-climacoef-distributionHowever, this can give us some illogical scenarios: for example, if the "uniform" distribution was selected, one day can have a climacoef of 0.1, but the next day can have a climacoef of 1.5 (remember that this affects the growth of grass, so it doesn't make sense that from one day to the next (within the same season), the climate changes in such a sudden way.
o	Moreover, it may not make much sense for the distribution to change from one season to another: it may not be logical to have a winter that follows a homogeneous distribution and then a spring with an "exponential_low" distribution.
o	Because of this, and after speaking with Irene, we decided to change this logic, so now the climacoef distribution will change seasonally, not daily, and the distribution will be annual, not seasonal.

2) Added a “Go (1 year)” button in the interface (The idea behind this is to give the observer the opportunity to choose the climate distribution for the next year).

3) Changed default value for the “DM-cm-ha” variable: from 1 to 180 (as in the base model, SequiaBasalto)