Changed "grass-height" equation: now "grass-height" is calculated using the option 1; FIGURE 2 IS REPLICATED!!!!!!!!
This change corresponds to what Francisco told me in his email.

In addition, with this change I was able to REPLICATE FIGURE 2 PERFECTLY, AND FIGURES 4, 5 AND 7 WERE IMPROVED (especially FIGURES 5 AND 7).

FIGURE 3 IS STILL NOT REPRODUCIBLE, BUT AFTER TELLING THIS TO FRANCISCO, HE SAID THAT HIS ORIGINAL FIGURE 3 IS NOT CORRECT, AND HE AGREED WITH ME THAT THE FIGURE 3 FROM THE NETLOGO MODEL IS CORRECT.