Added "Meat production (kg/ha)" monitor and plot; Changed criteria for how cows are sold during ordinary sales for the market-oriented and environmental-oriented farmer ;  Added document "FARMERS PROTOTYPES.docx"
Added "Meat production (kg/ha)" monitor and plot to the interface to measure the productivity of the livestock system.

Changed criteria for how cows are sold during ordinary sales for the market-oriented and environmental-oriented farmer (now they sell all old empty cows, regardless of the total number of cattle, and they will only sell empty heifers and cows if the number of cattle is above "keep-n-cattle").

Added document "FARMERS PROTOTYPES.docx" that explains the farmer behavior in the current version of the model.