Changed some "if = XXX" to "if >= XXX"
Changes have been made to "age = weaned-calf-age-min" and "lactating-time = lactation-period". Now they are "age >= weaned-calf-age-min" and "lactation-time >= lactation-period".

The logic behind these changes is that "As long as age remains an integer there is no problem, but the moment age becomes a real number (say using set days-per-tick 1.0) there is the possibility of floating points, and you get age becoming numbers like 10.00000001 and that would lead to not meeting the condition. A more robust implementation would be age >= weaned-calf-age-min"