Fixed a bug related to the maximum weight for the cow-with-calf age class
The maximum live weight of adult cows was limited to 650 kg, but I made a mistake in the code while implementing it: only cows with [cow?] were affected by this limit (and cows with [cow-with-cald?] were excluded, so they weren't affected by this limit).

Now, all adult cows (i.e., cows with [cows?] and [cow-with-calf?]) can grow to a maximum live weight of 650 kg.