1) Changed criteria for the extraordinary sales for the market farmer; 2) Updated "FARMER PROFILES" document
How extraordinary sales works: if at any moment of the year the average live weight of the herd is lower than the minimum live weight desired by the farmer, the farmer considers that the production of the system is at risk, triggering the sale of a certain number of animals until the average live weight of the herd returns to normal levels (i.e., is above the minimum live weight desired by the farmer) or until the system reaches the minimum herd size set by the farmer. Animals with the lowest OR highest live weight are sold following the next order: 1) male calves and steers, 2) old cows, 3) heifers and cows. In other words, if all males have been sold and the live weight of the herd is still below the threshold or the number of animals is still above the minimum herd size, the farmer will start selling old cows until the average live weight of the herd is above the threshold or the system reaches the minimum number of animals. If these criteria are not met even after selling all old cows, the farmer starts selling heifers and cows until one of these two criteria is met.