IMPORTANT CHANGES: 1) I SUBSTITUTED THE VARIABLE "gh-individual" for "grass-height" in the "grow-grass", "LWG" and "DM-consumption" procedures;  2) THE LOGIC OF THE ENVIRONMENTAL FARMER'S EXTRAORDINARY SALES HAS BEEN REDEFINED.; 3) THE ENVIRONMENTAL FARMER'S ROTATIONAL GRAZING LOGIC HAS BEEN REDEFINED.;  4) NOW FEED SUPPLEMENTATION WORKS THE SAME FOR THE MARKET FARMER AND THE ENVIRONMENTALFARMER.
1) I SUBSTITUTED THE VARIABLE "gh-individual" for "grass-height" in the "grow-grass", "LWG" and "DM-consumption" procedures.

Apparently, everything works fine (I remember that the reason I created a "gh-individual" variable in the past was because there was a problem with the livestock population: it was growing indefinitely....). With this change, it seems to make more sense when you look at the plots.

As a result, I deleted the procedure "kgDM/cow", used to calculate the "gh-individual" variable.

2) THE LOGIC OF THE ENVIRONMENTAL FARMER'S EXTRAORDINARY SALES HAS BEEN REDEFINED.

I have added an equation to determine the carrying capacity of the system (actual and estimated). To achieve this formula, I have based it on the concept of AUM (Animal Unit Month).

At the beginning of the season, the environmental farmer calculates the estimated-carrying-capacity (in Animal Units), so that if at any time during the season the number of AU (Animal Units) in the system exceeds the AU of the estimated-carrying-capacity, the farmer sells X animals until the AU equals the AU of the estimated-carrying-capacity.

3) THE ENVIRONMENTAL FARMER'S ROTATIONAL GRAZING LOGIC HAS BEEN REDEFINED.

Cows are now moved at the end of the season and the number of animals is adjusted to the capacity of the paddock through the extraordinary sales.

4) NOW FEED SUPPLEMENTATION WORKS THE SAME FOR THE MARKET FARMER AND THE ENVIRONMENTALFARMER.

As was the case before, the market farmer was more resilient in times of drought than the environmental farmer because of this strategy.