MODEL FINISHED!!!!!! LIST OF CHANGES HERE 1) CHANGED AGE CLASSES "MIN-WEIGHT" VALUE; 2) ABOUT THE EXTRAORDINARY SALES OF MARKET FARMER; 3) BUG FIXED; 4) ADDED "Market-fasb" TO FARMER PROFILES.
1) Changes min-weight value for each age class:
weaned-calves: from 60 to 100
heifers/steers: from 100 to 150
cows/cows-with-calves/pregnants/bulls: from 180 to 220

I made these changes to be consistent with the assumption that "1 AU (380 kg) has a minimum weight of 220 kg". 1 AU is an adult cow weighing 380 kg, and before this change the minimum weight for adult cows was 180 kg.

Since the minimum weight for adult cows has been increased to 220, the minimum weights for the other age classes have been increased accordingly.

2) For the market farmer's extraordinary sales, this farmer now takes into account the average live weight of the animals EXCLUDING born calves (previously born calves were also taken into account).

In addition, this change also makes it consistent with the exceptional sales of the environmental farmer, which takes into account the AU of all animals EXCEPT born calves.

3) Fixed a bug related to the "Daily Cost" plot in the interface that didn't show the "Supplement-costs".

4) To simplify and make the sensitivity analysis easier, I have added the strategy "Supplement-cows-before-breeding-season" within the "Farmer-Profile" chooser, with the name "market-fsb".