ADDED "environmental-fmincows" TO FARMER PROFILES
I have created an alternative version of the environmental farmer, which I have called "environmental-fmincows" (can be selected in the "farmer-profile" slider) and which takes up the original feed supplementation logic of the environmental farmer, in which he only buys supplementation when the system is below the minimum number of animals "when the system meets or falls below the minimum herd size desired by the farmer, the farmer purchases feed when the live weight of an animal is below a minimum live weight set by the farmer. This threshold is different for each age group. The amount of feed the farmer needs to give an animal in order for that animal to gain one kilogram of weight is determined by the Feed Conversion Ratio (FCR)."